Revert "Consolidate warrior code"
