+++ test/WarriorTagSystem.test.js
@@ -1,6 +1,9 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
-import { TagManager, CLASSES, UNITS, WARRIOR_SKILLS } from '../js/managers/warriormanager.js';
+import { TagManager } from '../js/managers/TagManager.js';
+import { CLASSES } from '../data/class.js';
+import { UNITS } from '../data/unit.js';
+import { WARRIOR_SKILLS } from '../data/warriorSkills.js';
 
 const tagManager = new TagManager({});
