+++ tests/unit/warriorSkillsAIUnitTests.js
@@ -1,6 +1,7 @@
 // tests/unit/warriorSkillsAIUnitTests.js
 
-import { WarriorSkillsAI, WARRIOR_SKILLS } from '../../js/managers/warriormanager.js';
+import { WarriorSkillsAI } from '../../js/managers/warriorSkillsAI.js';
+import { WARRIOR_SKILLS } from '../../data/warriorSkills.js';
 import { GAME_EVENTS, ATTACK_TYPES, GAME_DEBUG_MODE } from '../../js/constants.js';
 
 export function runWarriorSkillsAIUnitTests() {
